Improve email_handler
- remove getFilesToBeSent separate function, and include it in email_handler
- Attach all non .py folders in email to be sent
